SSL errors vague or misleading
When some SSL option like --key or --cert is wrong or missing SSL does
very vague messaging:

  Error connection to database: TLS/SSL error: No such file or directory
  Error connection to database: TLS/SSL error: The requested data were not available.
  Error connection to database: TLS/SSL error: Error while reading file.
  Error connection to database: TLS/SSL error: Base64 decoding error.

We can at least make this more understandable if the required option
is missing or if file not exists. The patch checks these criteria for
--ssl-mode=required (or preferred) for options --key and --cert and
for checking server identity mode it checks options --ca or --capath.

Bug description:

https://jira.mariadb.org/browse/CONTRIB-7