CONTRIB-14 mydumper dumps table for view
The logic in write_view_definition_into_file() of making duplicate
CREATE TABLE statements for views commented in the code as "we create
tables as workaround for view dependencies" pursues the goal of fixing
Bug#10927.

But this was done with undefined storage engine. Default storage
engine prohibits such amount of columns. We add ENGINE=MEMORY which
allows such amount of columns. Also we add TEMPORARY directive as it
was done in Bug#10927 fix.