Respect environment for connection
Client library kind of "supports" some of these environment variables,
but this support is buggy and incomplete. F.ex. mysqldump doesn't
respect MYSQL_HOST and setting MYSQL_TCP_PORT will not force it to use
TCP protocol. We treat this environment in upper layer so we can
guarantee and understand their work.

--debug enabled for any version. Tracing works ok on earlier versions.