CONTRIB-42 Lost row in dump for table with char-based PK
Char-based chunking algorithm in mydumper is missing first row because
of "<" in condition when comparing cmin_escaped. This comparison is
wrong for the very first chunk and leads to first row missed.

The fix includes first row in chunk and excludes last row unless last
row is cmax.

The problem was hidden in v0.15.2-6 because any chunking algorithm was
disabled due to the bug fixed in 640ece1 (myloader ignores
--max-threads-per-table). So v0.15.2-7 is affected by this bug since
it enables chunking algorithms.